API References

YaoBlocksQobj.convert_to_qbir — Method

convert_to_qbir(inst)

Converts Qobj based instructions back to YaoIR.

inst: The Qobj based instructions.

For Example:

q = convert_to_qobj(chain(1, put(1 => H))) 
ir = q.experiments[1].instructions |> convert_to_qbir

YaoBlocksQobj.create_experiment — Method

create_experiment(qc, exp_header, exp_config)

Returns and experiment type that consits of instructions.

qc: An Array of ChainBlock(circuits that are to be run).

exp_header(optional): Array of User-defined structure that contains metadata on each experiment and

is not used by the backend. The header will be passed through to the result data structure unchanged. For example, this may contain a fitting parameters for the experiment. In addition, this header can contain a mapping of backend memory and backend qubits to OpenQASM registers. This is because an OpenQASM circuit may contain multiple classical and quantum registers, but Qobj flattens them into a single memory and single set of qubits.

exp_config (optional): An Array of Configuration structure for user settings that can be different in each

experiment. These will override the configuration settings of the whole job.

YaoBlocksQobj.generate_inst — Method

generate_inst(qc)

Parses the YaoIR into a list of IBMQ supported instructions

qc: A ChainBlock(circuit that is to be run).